HLT-2423 : SimplesSearchService: upgrade tranIds to use Long type

+++ src/com/ctm/dao/health/HealthTransactionDao.java
@@ -137,7 +137,7 @@ public class HealthTransactionDao {
 	 * @param transactionId
 	 * @param errors value to write
 	 */
-	public void writeAllowableErrors(int transactionId , String errors) throws DaoException {
+	public void writeAllowableErrors(long transactionId , String errors) throws DaoException {
 
 		SimpleDatabaseConnection dbSource = new SimpleDatabaseConnection();
 		PreparedStatement stmt = null;
